pyproject.toml
[tool.poetry]
name = "tg_bank_forwarder"
version = "0.1.0"
description = ""
authors = ["Gonzalo Rizzo <me@gonzalorizzo.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
python-dotenv = "^0.21.0"
lxml = "^4.9.1"
pydantic = "^1.10.2"
pytelegrambotapi = {extras = ["aiohttp"], version = "^4.8.0"}
apscheduler = "^3.9.1.post1"
rich = "^12.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
